requirements.txt
PyYAML==3.11
praw==2.1.19
requests==2.5.0
six==1.8.0
update-checker==0.11
wsgiref==0.1.2
